+++ source/memory.h
@@ -8,6 +8,6 @@ void memset16(vu16 *dst, const u16 c, u16 hwsize);
 
 void rnd_seed(int seed);
 u16 rnd_random();
-s16 rnd_random_bounded(s16 min, s16 max);
+int rnd_random_bounded(int min, int max);
 
 #endif
